+++ IServices/IDetecter.cs
@@ -6,8 +6,8 @@ namespace BraimChallenge.IServices
     {
         public string[] HeaderData(string header);
         public int DetectUserAuth(string Authorize);
-        public int DetectId(int? accountId);
-        public int DetectAccount(int? accountId, string Authorize, List<Account> accountList);
+        public int DetectId(long? accountId);
+        public int DetectAccount(long? accountId, string Authorize, List<Account> accountList);
 
     }
 }
